requirements.txt
cassandra-driver==3.25.0
category-encoders==2.2.2
certifi==2021.5.30
click 
cloudpickle==1.6.0
colorama==0.4.4
dnspython==1.16.0
email-validator 
Flask==2.0.1
Flask-WTF 
gunicorn==20.1.0
idna 
importlib-metadata 
itsdangerous 
Jinja2 
joblib==1.0.1
llvmlite==0.36.0
MarkupSafe 
numba==0.53.1
numpy 
pandas==1.3.2
patsy==0.5.1
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
shap==0.39.0
six 
slicer==0.0.7
statsmodels==0.12.2
threadpoolctl==2.2.0
tqdm==4.62.1
Werkzeug==2.0.1
wincertstore==0.2
WTForms 
zipp 
